feat(statistics): add category icons to dynamics chart hover tooltip
The tooltip (unlike the Y-axis tick from the comparison chart) renders
as regular DOM content, not inside recharts' SVG space, so it can
reuse NameWithOptionalIcon directly. Wired via tooltipProps.content,
same override mechanism already used for the legend.
